Convert Force-Velocity profile back to Acceleration-Velocity profile

This function converts back the Force-Velocity profile (FVP; F0 and V0 parameters) to Acceleration-Velocity profile (AVP; MSS and MAC parameters)

Usage

convert_FVP(
  F0,
  V0,
  bodymass = 75,
  inertia = 0,
  resistance = 0,
  wind_velocity = 0,
  ...
)

Arguments

F0, V0

Numeric vectors. FV profile parameters

bodymass

Body mass in kg. Used to calculate relative power and forwarded to get_air_resistance

inertia

External inertia in kg (for example a weight vest, or a sled). Not included in the air resistance calculation

resistance

External horizontal resistance in Newtons (for example tether device or a sled friction resistance)

wind_velocity

In meters per second (m/s). Use negative number as head wind, and positive number as back wind. Default is 0m/s (no wind)

...

Forwarded to predict_power_at_distance

Value

A list with calculated MSS and MAC parameters

Examples

FVP <- create_FVP(7, 8.3, inertia = 10, resistance = 50)

convert_FVP(FVP$F0, FVP$V0, inertia = 10, resistance = 50)
#> $MSS
#> [1] 7
#> 
#> $MAC
#> [1] 8.3
#>