Function that finds the distance at which the sprint, probe, or FV profile is optimal (i.e., equal to 100 perc)

Function that finds the distance at which the sprint, probe, or FV profile is optimal (i.e., equal to 100 perc)

Usage

find_optimal_distance(..., optimal_func = optimal_FV, min = 1, max = 60)

Arguments

...

Forwarded to selected optimal_func

optimal_func

Selected profile optimization function. Default is optimal_FV

min, max

Distance over which to find optimal profile distance

Value

Distance

Examples

MSS <- 10
MAC <- 8
bodymass <- 75

fv <- create_FVP(MSS, MAC, bodymass)

find_optimal_distance(
  F0 = fv$F0,
  V0 = fv$V0,
  bodymass = fv$bodymass,
  optimal_func = optimal_FV,
  method = "max"
)
#> [1] 18.80181

find_optimal_distance(
  MSS = MSS,
  MAC = MAC,
  optimal_func = optimal_MSS_MAC
)
#> [1] 15.8214

find_optimal_distance(
  MSS = MSS,
  MAC = MAC,
  optimal_func = probe_MSS_MAC
)
#> [1] 16.01597