Get Force-Velocity Profile

Provides Force-Velocity (FV) profile modified using ideas by Pierre Samozino and JB-Morin, et al. (2016) and Pierre Samozino and Nicolas Peyror, et al (2021).

make_FV_profile(MSS, MAC, bodymass = 75, inertia = 0, resistance = 0, ...)

Arguments

MSS, MAC

Numeric vectors. Model parameters

bodymass

Body mass in kg. Used to calculate relative power and forwarded to get_air_resistance

inertia

External inertia in kg (for example a weight vest, or a sled). Not included in the air resistance calculation

resistance

External horizontal resistance in Newtons (for example tether device or a sled friction resistance)

...

Forwarded to get_air_resistance for the purpose of calculation of air resistance and power

Value

List containing the following elements:

bodymass

Returned bodymass used in FV profiling

F0

Horizontal force when velocity=0

F0_rel

F0 divided by bodymass

V0

Velocity when horizontal force=0

Pmax

Maximal horizontal power

Pmax_rel

Pmax divided by bodymass

FV_slope

Slope of the FV profile. See References for more info

References

Samozino P, Rabita G, Dorel S, Slawinski J, Peyrot N, Saez de Villarreal E, Morin J-B. 2016. A simple method for measuring power, force, velocity properties, and mechanical effectiveness in sprint running: Simple method to compute sprint mechanics. Scandinavian Journal of Medicine & Science in Sports 26:648–658. DOI: 10.1111/sms.12490.

Samozino P, Peyrot N, Edouard P, Nagahara R, Jimenez‐Reyes P, Vanwanseele B, Morin J. 2022. Optimal mechanical force‐velocity profile for sprint acceleration performance. Scandinavian Journal of Medicine & Science in Sports 32:559–575. DOI: 10.1111/sms.14097.

Examples

data("jb_morin")

m1 <- model_radar_gun(time = jb_morin$time, velocity = jb_morin$velocity)

fv_profile <- make_FV_profile(
  MSS = m1$parameters$MSS,
  MAC = m1$parameters$MAC,
  bodyheight = 1.72,
  bodymass = 120,

)

fv_profile
#> $bodymass
#> [1] 120
#> 
#> $F0
#> [1] 829.9911
#> 
#> $F0_rel
#> [1] 6.916593
#> 
#> $V0
#> [1] 9.534875
#> 
#> $Pmax
#> [1] 1978.465
#> 
#> $Pmax_rel
#> [1] 16.48721
#> 
#> $FV_slope
#> [1] -0.7253995
#>