Mixed Model Using Instantaneous Velocity

This function models the sprint instantaneous velocity using mono-exponential equation and non-linear mixed model using nlme to estimate fixed and random maximum sprinting speed (MSS) and relative acceleration (TAU) parameters. In mixed model, fixed and random effects are estimated for MSS and TAU parameters using athlete as levels. velocity is used as target or outcome variable, and time as predictor.

mixed_model_using_instant_velocity(
  data,
  time,
  velocity,
  athlete,
  time_delay,
  na.rm = FALSE
)

Arguments

data
Data frame

time
Character string. Name of the column in data

velocity
Character string. Name of the column in data

athlete
Character string. Name of the column in data. Used as levels in the nlme

time_delay
Character string. Name of the column in data. Used to filter out noisy data from the radar gun. Default is 0. See more in Samozino (2018)

na.rm
Logical. Default is FALSE

Value

List object with the following elements:

parameters

List with two data frames: fixed and random containing the following estimated parameters: MSS, TAU, MAC, and PMAX

model_fit

List with the following components: RSE, R_squared, minErr, maxErr, maxAbsErr, and RMSE

model

Model returned by the nlme function

data

Data frame used to estimate the sprint parameters, consisting of athlete, time, velocity, time_delay, and pred_velocity columns

References

Samozino P. 2018. A Simple Method for Measuring Force, Velocity and Power Capabilities and Mechanical Effectiveness During Sprint Running. In: Morin J-B, Samozino P eds. Biomechanics of Training and Testing. Cham: Springer International Publishing, 237–267. DOI: 10.1007/978-3-319-05633-3_11.

Examples

data("radar_gun_data")
mixed_model <- mixed_model_using_instant_velocity(radar_gun_data, "time", "velocity", "athlete")
mixed_model$parameters

#> $fixed
#>        MSS       TAU      MAC      PMAX
#> 1 8.240711 0.4805574 17.14824 0.9900337
#> 
#> $random
#>     athlete      MSS       TAU       MAC      PMAX
#> 1      John 7.998751 0.6021399 13.283876 1.2040918
#> 2 Kimberley 7.205054 0.3003699 23.987266 0.5410454
#> 3       Jim 8.997225 1.0995085  8.182951 2.4731312
#> 4     James 9.999072 0.2003679 49.903555 0.5008733
#> 5  Samantha 7.003456 0.2004006 34.947270 0.3508743
#>