Mixed Model Using Instantaneous Velocity

This function models the sprint instantaneous velocity using mono-exponential equation and non-linear mixed model using nlme to estimate fixed and random maximum sprinting speed (MSS) and relative acceleration (TAU) parameters. In mixed model, fixed and random effects are estimated for MSS and TAU parameters using athlete as levels. velocity is used as target or outcome variable, and time as predictor.

mixed_model_using_radar(
  data,
  time,
  velocity,
  athlete,
  time_correction = 0,
  na.rm = FALSE
)

Arguments

data
Data frame

time
Character string. Name of the column in data

velocity
Character string. Name of the column in data

athlete
Character string. Name of the column in data. Used as levels in the nlme

time_correction
Numeric vector. Used to filter out noisy data from the radar gun. This correction is done by adding time_correction to time. Default is 0. See more in Samozino (2018)

na.rm
Logical. Default is FALSE

Value

List object with the following elements:

parameters

List with two data frames: fixed and random containing the following estimated parameters: MSS, TAU, MAC, and PMAX

model_fit

List with the following components: RSE, R_squared, minErr, maxErr, and RMSE

model

Model returned by the nlme function

data

Data frame used to estimate the sprint parameters, consisting of athlete, time, velocity, and pred_velocity columns

References

Samozino P. 2018. A Simple Method for Measuring Force, Velocity and Power Capabilities and Mechanical Effectiveness During Sprint Running. In: Morin J-B, Samozino P eds. Biomechanics of Training and Testing. Cham: Springer International Publishing, 237–267. DOI: 10.1007/978-3-319-05633-3_11.

Examples

data("radar_gun_data")
mixed_model <- mixed_model_using_radar(radar_gun_data, "time", "velocity", "athlete")

#mixed_model$parameters
coef(mixed_model)

#> $fixed
#>                 MSS                 TAU                 MAC                PMAX 
#>            8.301178            1.007782            8.237080           17.094367 
#>     time_correction distance_correction 
#>            0.000000            0.000000 
#> 
#> $random
#>     athlete      MSS       TAU      MAC     PMAX time_correction
#> 1     James 9.998556 1.1108457 9.000851 22.49888               0
#> 2       Jim 7.997945 0.8886712 8.999892 17.99516               0
#> 3      John 8.000051 1.0690357 7.483427 14.96695               0
#> 4 Kimberley 9.005500 1.2855706 7.005061 15.77102               0
#> 5  Samantha 6.503839 0.6847851 9.497635 15.44277               0
#>   distance_correction
#> 1                   0
#> 2                   0
#> 3                   0
#> 4                   0
#> 5                   0
#>