Model Using Split Times

This function models the sprint split times using mono-exponential equation that estimates maximum sprinting speed (MSS) and relative acceleration (TAU). time is used as target or outcome variable, and distance as predictor.

model_using_split_times(distance, time, weights = 1, na.rm = FALSE)

Arguments

distance
Numeric vector

time
Numeric vector

weights
Numeric vector. Default is vector of 1 This is used to give more weight to particular observations. For example, use 1\distance to give more weight to observations from shorter distances.

na.rm
Logical. Default is FALSE

Value

List object with the following elements:

parameters

List with the following estimated parameters: MSS, TAU, MAC, and PMAX

model_fit

List with the following components: RSE, R_squared, minErr, maxErr, maxAbsErr, and RMSE

model

Model returned by the nls function

data

Data frame used to estimate the sprint parameters, consisting of distance, time, weights and pred_time columns

Examples

split_times <- data.frame(
distance = c(5, 10, 20, 30, 35),
time = c(1.21, 1.99, 3.38, 4.71, 5.30)
)

sprint_model <- with(
  split_times,
  model_using_split_times(distance, time)
)

sprint_model$parameters

#> $MSS
#> [1] 7.553185
#> 
#> $TAU
#> [1] 0.706663
#> 
#> $MAC
#> [1] 10.68852
#> 
#> $PMAX
#> [1] 20.1831
#>