Kinematics prediction functions

Predicts kinematic from known MSS and TAU parameters

predict_velocity_at_time(time, MSS, TAU, time_correction = 0)

predict_distance_at_time(
  time,
  MSS,
  TAU,
  time_correction = 0,
  distance_correction = 0
)

predict_acceleration_at_time(time, MSS, TAU, time_correction = 0)

predict_time_at_distance(
  distance,
  MSS,
  TAU,
  time_correction = 0,
  distance_correction = 0
)

predict_velocity_at_distance(
  distance,
  MSS,
  TAU,
  time_correction = 0,
  distance_correction = 0
)

predict_acceleration_at_distance(
  distance,
  MSS,
  TAU,
  time_correction = 0,
  distance_correction = 0
)

predict_acceleration_at_velocity(velocity, MSS, TAU)

predict_relative_power_at_distance(
  distance,
  MSS,
  TAU,
  time_correction = 0,
  distance_correction = 0
)

predict_relative_power_at_time(time, MSS, TAU, time_correction = 0)

predict_kinematics(object, max_time = 6, frequency = 100)

Arguments

time, distance, velocity
Numeric vectors

MSS, TAU
Numeric vectors. Model parameters

time_correction
Numeric vector. Used for correction. Default is 0. See references for more info

distance_correction
Numeric vector. Used for correction. Default is 0. See vignettes for more info

object
shorts_model or shorts_mixed_model object

max_time
Predict from 0 to max_time. Default is 6seconds

frequency
Number of samples within one second. Default is 100Hz

Value

Numeric vector

Data frame

References

Haugen TA, Tønnessen E, Seiler SK. 2012. The Difference Is in the Start: Impact of Timing and Start Procedure on Sprint Running Performance: Journal of Strength and Conditioning Research 26:473–479. DOI: 10.1519/JSC.0b013e318226030b.

Samozino P. 2018. A Simple Method for Measuring Force, Velocity and Power Capabilities and Mechanical Effectiveness During Sprint Running. In: Morin J-B, Samozino P eds. Biomechanics of Training and Testing. Cham: Springer International Publishing, 237–267. DOI: 10.1007/978-3-319-05633-3_11.

Examples

MSS <- 8
TAU <- 0.7

time_seq <- seq(0, 6, length.out = 10)

df <- data.frame(
  time = time_seq,
  distance_at_time = predict_distance_at_time(time_seq, MSS, TAU),
  velocity_at_time = predict_velocity_at_time(time_seq, MSS, TAU),
  acceleration_at_time = predict_acceleration_at_time(time_seq, MSS, TAU)
)

df$time_at_distance <- predict_time_at_distance(df$distance_at_time, MSS, TAU)
df$velocity_at_distance <- predict_velocity_at_distance(df$distance_at_time, MSS, TAU)
df$acceleration_at_distance <- predict_acceleration_at_distance(df$distance_at_time, MSS, TAU)
df$acceleration_at_velocity <- predict_acceleration_at_velocity(df$velocity_at_time, MSS, TAU)

df

#>         time distance_at_time velocity_at_time acceleration_at_time
#> 1  0.0000000         0.000000         0.000000         11.428571429
#> 2  0.6666667         1.893933         4.913430          4.409386364
#> 3  1.3333333         5.900272         6.809135          1.701235209
#> 4  2.0000000        10.721623         7.540539          0.656372792
#> 5  2.6666667        15.857422         7.822730          0.253242608
#> 6  3.3333333        21.114543         7.931606          0.097706394
#> 7  4.0000000        26.418472         7.973612          0.037697209
#> 8  4.6666667        31.740460         7.989819          0.014544386
#> 9  5.3333333        37.069416         7.996072          0.005611534
#> 10 6.0000000        42.401061         7.998484          0.002165049
#>    time_at_distance velocity_at_distance acceleration_at_distance
#> 1                 0                    0             11.42857....
#> 2      0.666666....         4.913429....             4.409386....
#> 3      1.333333....         6.809135....             1.701235....
#> 4                 2         7.540539....             0.656372....
#> 5      2.666666....         7.822730....             0.253242....
#> 6      3.333333....         7.931605....             0.097706....
#> 7                 4         7.973611....             0.037697....
#> 8      4.666666....         7.989818....             0.014544....
#> 9      5.333333....         7.996071....             0.005611....
#> 10                6         7.998484....             0.002165....
#>    acceleration_at_velocity
#> 1              11.428571429
#> 2               4.409386364
#> 3               1.701235209
#> 4               0.656372792
#> 5               0.253242608
#> 6               0.097706394
#> 7               0.037697209
#> 8               0.014544386
#> 9               0.005611534
#> 10              0.002165049